optional modification

+++ lib/src/models/wind.dart
@@ -6,10 +6,10 @@ part 'wind.g.dart';
 @JsonSerializable()
 class Wind extends Equatable {
   @JsonKey(name: 'speed')
-  final int speed;
+  final double speed;
 
   @JsonKey(name: 'deg')
-  final int deg;
+  final double deg;
 
   const Wind({
     required this.speed,
